src/test/java/programmers/courses30/lesson17680/SolutionTest.java
package programmers.courses30.lesson17680;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class SolutionTest {

    @Test
    void solution() {
        assertEquals(50, new Solution().solution(3, new String[]{"Jeju", "Pangyo", "Seoul", "NewYork", "LA", "Jeju", "Pangyo", "Seoul", "NewYork", "LA"}));
        assertEquals(21, new Solution().solution(3, new String[]{"Jeju", "Pangyo", "Seoul", "Jeju", "Pangyo", "Seoul", "Jeju", "Pangyo", "Seoul"}));
        assertEquals(60, new Solution().solution(2, new String[]{"Jeju", "Pangyo", "Seoul", "NewYork", "LA", "SanFrancisco", "Seoul", "Rome", "Paris", "Jeju", "NewYork", "Rome"}));
        assertEquals(52, new Solution().solution(5, new String[]{"Jeju", "Pangyo", "Seoul", "NewYork", "LA", "SanFrancisco", "Seoul", "Rome", "Paris", "Jeju", "NewYork", "Rome"}));
        assertEquals(16, new Solution().solution(2, new String[]{"Jeju", "Pangyo", "NewYork", "newyork"}));
        assertEquals(25, new Solution().solution(0, new String[]{"Jeju", "Pangyo", "Seoul", "NewYork", "LA"}));
        assertEquals(25, new Solution().solution(0, new String[]{"Jeju", "Jeju", "Jeju", "Jeju", "Jeju"}));
    }
}